+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file.
 ### Improvements
 - Upgrade to fido2 1.1.0 and simplify hc.lib.webauthn
 - Add handling for ipv4address:port values in the X-Forwarded-For header (#714)
+- Add a form for submitting Signal CAPTCHA solutions
 
 ## v2.4.1 - 2022-10-18